Refactoring admin (#1222)
* refactored two wiki created files

* refactored two wiki created files

* update codecimate yml file

* update codecimate yml file

* refactoring pages and prop types

* imports update

* reducing pages to meet method line of 200

* reducing pages to meet method line of 200

* updating the code climate file

* wiki created action bar and footer

* wiki created action bar and footer

* exported types

* table column refactor

* table col date refactor

* table col date refactor

* refacoting functions

* refactoring functions in WikiInsight Page

* Line

* refactored sort func

* refactored sort func

* increased method lines in config

* Update src/components/Admin/WikiCreatedInsight/GetWikisComponent.tsx


* Update src/components/Admin/WikiCreatedInsight/WikiCreatedActionBar.tsx


* fixed changes requested

* exported hide notifocation prop type

* Update src/components/Admin/WikiEditorInsight/InsightTableWikiEditors.tsx


* Update src/components/Admin/WikiEditorInsight/InsightTableWikiEditors.tsx


* Update src/components/Admin/WikiEditorInsight/InsightTableWikiEditors.tsx


* Update src/components/Admin/WikiEditorInsight/InsightTableWikiEditors.tsx


* Update src/components/Admin/WikiCreatedInsight/PromotedWikiContent.tsx


* requested changes done

* more fixes from changes requested

* more fixes from changes requested

* more fixes from changes requested

* naming Colors

* naming Colors

* fixing more colors name

* fixing more colors name

* solving issue of named colors on graph

---------
