+++ src/components/Admin/GraphHeads.tsx
@@ -3,7 +3,7 @@ import { Th, Text } from '@chakra-ui/react'
 
 export const TableHead = ({ text }: { text: string }) => {
   return (
-    <Th color="#718096" textTransform="capitalize" fontWeight="normal">
+    <Th color="primaryGray" textTransform="capitalize" fontWeight="normal">
       <Text fontWeight="bold">{text}</Text>
     </Th>
   )
