+++ src/components/Admin/WikiCreatedInsight/WikiCreatedActionBar.tsx
@@ -37,8 +37,7 @@ export const WikiCreatedActionBar = (props: WikiCreatedActionBarProps) => {
     onToggle,
   } = props
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const ApplyFilterItems = (e: any) => {
+  const ApplyFilterItems = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault()
     // get all checkboxes from form
     const checkboxes = Array.from(
@@ -91,7 +90,7 @@ export const WikiCreatedActionBar = (props: WikiCreatedActionBarProps) => {
         <Flex gap={5}>
           <InputGroup w="100%">
             <InputLeftElement pointerEvents="none">
-              <FiSearch color="#667085" />
+              <FiSearch color="metalGray" />
             </InputLeftElement>
             <Input
               type="text"
@@ -108,8 +107,7 @@ export const WikiCreatedActionBar = (props: WikiCreatedActionBarProps) => {
               _expanded={{ bg: 'brand.500', color: 'white' }}
             >
               <Button
-                borderColor="#E2E8F0"
-                _dark={{ borderColor: '#2c323d' }}
+                borderColor="wikiSortBorder"
                 py={2}
                 px={5}
                 leftIcon={sortIcon}
