+++ src/components/Admin/WikiCreatedInsight/WikiElemets.tsx
@@ -1,4 +1,4 @@
-import { WikiCreatedFooterProps, Wikis } from '@/types/admin'
+import { WikiCreatedFooterProps, WikiTableColProps } from '@/types/admin'
 import { ArrowBackIcon, ArrowForwardIcon } from '@chakra-ui/icons'
 import { Flex, Button, HStack, Icon, Text } from '@chakra-ui/react'
 import React from 'react'
@@ -21,7 +21,6 @@ export const WikiCreatedFooter = (props: WikiCreatedFooterProps) => {
       wikis && wikis?.length >= 10 && setPaginateOffset(paginateOffset + 10)
     )
   }
-
   const decreasePagination = () => {
     return (
       wikis && wikis?.length >= 10 && setPaginateOffset(paginateOffset - 10)
@@ -62,13 +61,6 @@ export const WikiCreatedFooter = (props: WikiCreatedFooterProps) => {
   )
 }
 
-type WikiTableColProps = {
-  item: Wikis
-  PromoteClickOne: () => void
-  PromoteClickTwo: () => void
-  ArchiveClickOne: () => void
-  ArchiveClickTwo: () => void
-}
 export const WikiTableCol = (props: WikiTableColProps) => {
   const {
     item,
@@ -82,9 +74,9 @@ export const WikiTableCol = (props: WikiTableColProps) => {
       <HStack spacing={5}>
         {!item.promoted ? (
           <Text
-            color={item.hidden ? '#E2E8F0' : '#FF5CAA'}
+            color={item.hidden ? 'tetiaryGray' : 'brand.500'}
             _dark={{
-              color: item.hidden ? '#51565F' : '#FF1A88',
+              color: item.hidden ? 'davyGray' : 'brand.800',
             }}
             cursor={item.hidden ? 'not-allowed' : 'pointer'}
             fontWeight="semibold"
@@ -95,9 +87,9 @@ export const WikiTableCol = (props: WikiTableColProps) => {
         ) : (
           <HStack spacing={2} onClick={PromoteClickTwo}>
             <Text
-              color="#E2E8F0"
+              color="tetiaryGray"
               _dark={{
-                color: item.hidden ? '#51565F' : '#495a68',
+                color: item.hidden ? 'davyGray' : 'denceGray',
               }}
               cursor={item.hidden ? 'not-allowed' : 'pointer'}
             >
@@ -106,12 +98,11 @@ export const WikiTableCol = (props: WikiTableColProps) => {
             <Icon
               fontSize="20px"
               cursor="pointer"
-              color="#F11a82"
+              color="electricPink"
               as={RiQuestionLine}
             />
           </HStack>
         )}
-
         {!item.hidden ? (
           <Text
             cursor={item.hidden ? 'not-allowed' : 'pointer'}
@@ -122,13 +113,13 @@ export const WikiTableCol = (props: WikiTableColProps) => {
           </Text>
         ) : (
           <HStack spacing={2} onClick={ArchiveClickTwo}>
-            <Text color="#F11a82" _dark={{ color: '#FF1A88' }} cursor="pointer">
+            <Text color="unArchiveWiki" cursor="pointer">
               Unarchive
             </Text>
             <Icon
               fontSize="20px"
               cursor="pointer"
-              color="#F11a82"
+              color="electricPink"
               as={RiQuestionLine}
             />
           </HStack>
@@ -141,7 +132,7 @@ export const WikiTableCol = (props: WikiTableColProps) => {
 export const WikiColDate = ({ colDate }: { colDate: string | undefined }) => {
   return (
     <HStack>
-      <Text color="#718096">
+      <Text color="primaryGray">
         {colDate
           ? new Date(colDate).toLocaleDateString('en-US', {
               year: 'numeric',
@@ -158,7 +149,7 @@ export const WikiColDate = ({ colDate }: { colDate: string | undefined }) => {
         as={BsDot}
         justifySelf="center"
       />
-      <Text color="#718096" textTransform="lowercase">
+      <Text color="primaryGray" textTransform="lowercase">
         {colDate
           ? new Date(colDate).toLocaleTimeString('en-US', {
               hour: 'numeric',
