+++ src/theme/foundations/colors.ts
@@ -25,4 +25,14 @@ export const colors = {
   primaryPink: '#E53E3E',
   tetiaryDark: '#1A202C',
   primaryGray: '#718096',
+  davyGray: '#51565F',
+  electricPink: '#f11a82',
+  taupeGray: '#898787',
+  mediumGreen: '#38A169',
+  dewGray: '#F0FFF4',
+  tangRed: '#DD6B20',
+  dazBlue: '#385C8A',
+  denceGray: '#495a68',
+  metalGray: '#667085',
+  rosePink: '#FF5DAA',
 }
